requirements.txt
PyYAML==6.0
hvac==0.8.2
pyjwt==2.3.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
requests==2.27.1
urllib3==1.26.9

kubernetes==22.6.0
ujson==5.4.0
pytest==7.1.2
pydantic==1.10.2
prettytable==2.5.0
wcwidth==0.2.5
